Fix: Move Language Settings to General Settings
