+++ src/components/settings/index.ts
@@ -1,4 +1,3 @@
 export * from './ApiManagement';
 export * from './ChatSettings';
-export * from './LanguageSettings';
 export * from './GeneralSettings';
